+++ docs/changelog.md
@@ -4,6 +4,11 @@
 
 ### Features
 
+* New metrics ([tpo/onion-services/onionprobe#78][]):
+    * From the outer descriptor wrapper:
+        * `descriptor-lifetime`.
+        * `revision-counter`.
+
 * Enhanced Grafana Dashboard ([tpo/onion-services/onionprobe#80][]).
 
 * Added debug messages for HS_DESC events:
@@ -11,6 +16,7 @@
 
   * HSDir used.
 
+[tpo/onion-services/onionprobe#78]: https://gitlab.torproject.org/tpo/onion-services/onionprobe/-/issues/78
 [tpo/onion-services/onionprobe#80]: https://gitlab.torproject.org/tpo/onion-services/onionprobe/-/issues/80
 
 ### Fixes